Add privacy notice
Change made on advice about what is needed to ensure compliance with new GDPR regulations.

Google analytics tracking is set in the settings.php file, stored locally on the server. I have made the required change on the production server - the change can be verified in the <head> section of the source of files on http://validator.iatistandard.org